requirements.txt
h5py==3.1.0
matplotlib==3.3.3
numpy==1.20.3
opencv_python==4.5.2.54
Pillow==9.1.1
pycocotools==2.0.2
pydantic==1.8.2
PyYAML==6.0
scalabel==0.3.0
scikit_image==0.18.3
tqdm==4.38.0
